Initial rework of place screen
